+++ src/database/migrations/2020_01_20_211539_region_c.php
@@ -11,24 +11,26 @@ class RegionC extends Migration {
 	 * @return void
 	 */
 	public function up() {
-		Schema::create('regions', function (Blueprint $table) {
-			$table->increments('id');
-			$table->unsignedInteger('company_id');
-			$table->string('name', 191);
-			$table->string('code', 4);
-			$table->unsignedInteger('state_id');
-			$table->unsignedInteger('created_by_id')->nullable();
-			$table->unsignedInteger('updated_by_id')->nullable();
-			$table->unsignedInteger('deleted_by_id')->nullable();
-			$table->timestamps();
-			$table->softdeletes();
+		if (!Schema::hasTable('regions')) {
+			Schema::create('regions', function (Blueprint $table) {
+				$table->increments('id');
+				$table->unsignedInteger('company_id');
+				$table->string('name', 191);
+				$table->string('code', 4);
+				$table->unsignedInteger('state_id');
+				$table->unsignedInteger('created_by_id')->nullable();
+				$table->unsignedInteger('updated_by_id')->nullable();
+				$table->unsignedInteger('deleted_by_id')->nullable();
+				$table->timestamps();
+				$table->softdeletes();
 
-			$table->foreign('state_id')->references('id')->on('states')->onDelete('CASCADE')->onUpdate('cascade');
+				$table->foreign('state_id')->references('id')->on('states')->onDelete('CASCADE')->onUpdate('cascade');
 
-			$table->unique(["code", "state_id"]);
-			$table->unique(["name", "state_id"]);
+				$table->unique(["code", "state_id"]);
+				$table->unique(["name", "state_id"]);
 
-		});
+			});
+		}
 
 	}
